🐛 fix(cli): move verify vocabulary to @lobechat/const/verify
`apps/cli` is a member of the isolated desktop workspace, where
`@lobechat/types` resolves to the hand-written stub in
`apps/desktop/stubs/types`. Value imports from it are therefore
unreachable, and #17166 added three to the CLI's verify command,
breaking `@lobehub/cli`'s build on canary.

Move the verify closed sets (verifier types, statuses, verdicts,
surfaces, evidence media), `normalizeVerifySurface`,
`DEFAULT_MAX_REPAIR_ROUNDS` and `VerifyRunOrigin` into a new zero-import
`@lobechat/const/verify` narrow subpath, which every workspace can
resolve. `@lobechat/types` re-exports the derived types, so type-only
consumers are untouched and it stays the one-stop domain vocabulary.

The whole vocabulary moves, not just the two symbols the CLI happened to
need — splitting on what one consumer imports is an accident, not a
boundary.
